% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr.R
\name{add_attr}
\alias{add_attr}
\title{Add an attribute}
\usage{
add_attr(tree, attr, vals)
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{attr}{The name of the new attribute, a character string.
It should not contain the underscore character \code{"_"}.}

\item{vals}{The values of the new attribute,
a vector of numeric values, logical values or character strings.
Length of the vector must be equal to the number of all cells in the \code{tree},
as returned from \code{\link{get_cells}}.}
}
\value{
The updated tree with the new attribute \code{attr} added, an object of class \code{"igraph"}.
}
\description{
Adds an attribute to a lineage or division tree.
}
