% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_growth_fit_pars.R
\name{add_attr_growth_fit_pars}
\alias{add_attr_growth_fit_pars}
\title{Compute growth curves of an attribute}
\usage{
add_attr_growth_fit_pars(LT, DT, attr, model = c("lin", "exp"), frameR)
}
\arguments{
\item{LT}{The lineage tree, an object of class \code{"igraph"}.}

\item{DT}{The corresponding division tree of the \code{LT}, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{LT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"}, \code{"frame"} and \code{"age"}.}

\item{model}{A character string naming the growth model to be fitted:
\itemize{
\item \code{"lin"} for fitting a \emph{linear} model
\code{\ifelse{html}{\out{y = at + b}}{\eqn{y = a \cdot t + b}}}
using the \emph{linear least squares} method provided by \code{\link[stats]{lm}}
\item \code{"exp"} for fitting an \emph{exponential} model
\code{\ifelse{html}{\out{y = y<sub>0</sub> e<sup>kt</sup>}}{\eqn{y = y_{0} \cdot e^{kt}}}}
using the \emph{non-linear least squares} method provided by \code{\link[stats]{nls}}
}
where \code{y} represents the \code{attr} and \code{t} the time in \emph{hours} (starting from \code{0}).}

\item{frameR}{Frame rate of the movie in \emph{frames} per \emph{minute}, a non-zero positive numeric value.}
}
\value{
The updated \code{DT} with the new attributes added, an object of class \code{"igraph"}.
}
\description{
Computes the growth curve of a numeric attribute for each cell in a division tree,
by fitting a \emph{linear} or \emph{exponential} model to the its attribute's time-series.
}
\details{
The estimated parameters as well as the RMSE of the fitted growth model
are added as attributes to the \code{DT}:
\itemize{
\item When \code{model = "lin"}:
\itemize{
\item \code{"<attr>_a"}, a non-zero positive numeric value in units of \code{attr} per \emph{hour}
\item \code{"<attr>_b"}, a positive numeric value in units of \code{attr}
\item \code{"<attr>_linRMSE"}
}
\item When \code{model = "exp"}:
\itemize{
\item \code{"<attr>_k"}, a non-zero positive numeric value in units of \code{attr} per \emph{hour}
\item \code{"<attr>_0"}, a non-zero positive numeric value in units of \code{attr}
\item \code{"<attr>_expRMSE"}
}
}
\code{NA} values are stored for cells that failed to fit the selected model
as well as for cells that are not included in the analysis, as returned from \code{\link{get_cells}}.
Information messages for cells that failed to fit the model are printed on the screen.
}
