% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_branch.R
\name{add_branch}
\alias{add_branch}
\title{Add a branch to an LT}
\usage{
add_branch(LT, branch, cell, cell_list, col_list, Ncols)
}
\arguments{
\item{LT}{The lineage tree to which the \code{branch} will be added, an object of class \code{"igraph"}.}

\item{branch}{The connected lineage tree (motherless branch) which will be added as a branch to the \code{LT}, an object of class \code{"igraph"}.}

\item{cell}{The label of the cell in the \code{LT} where the \code{branch} will be added, a character string.
It can be any valid candidate mother cell, as returned from \code{\link{get_cand_mother_cells}}.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}
}
\value{
The new LT with the branch added, an object of class \code{"igraph"}.
}
\description{
Adds a lineage tree as a branch to another lineage tree at a specific position,
defined by the cell which will be the mother of the root of the branch.
}
\section{Prerequisites}{

See the \emph{Prerequisites} section of \code{\link{get_cand_mother_cells}}.
}

\seealso{
\code{\link{extract_branch}} for extracting a branch from a tree,
\code{\link{isConnected}} for checking if a tree is connected.
}
