% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_life.R
\name{create_cell_life}
\alias{create_cell_life}
\title{Visualize cell life}
\usage{
create_cell_life(DT, cells = "all", cell_list, col_list, Ncols,
  saveT = c("sep", "comb"), savePath = getwd())
}
\arguments{
\item{DT}{The division tree where the cells specified in \code{cells} belong, an object of class \code{"igraph"}.}

\item{cells}{The labels of the cells in the \code{DT} whose life images will be created,
a vector of character strings.
They can be any non-root cells, as returned from \code{\link{get_cells}}.
The default value \code{"all"} stands for all non-root cells.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{saveT}{A character string naming the options of image saving:
\itemize{
\item \code{"sep"} saves a separate \code{.png} file for each frame of the cell's lifespan,
named as \code{"<cell>_f<first_frame>_<i>.png"}.
Image files for each cell will be saved in a directory named after the cell.
\item \code{"comb"} saves a single \code{.png} file with all frames of the cell's lifespan combined
in \emph{left-to-right, top-to-bottom} order, named as \code{"<cell>_f<first_frame>.png"}.
}
\code{"<cell>"} is the label of the cell, \code{"<first_frame>"} is its birth frame and
\code{"<i>"} is the instant of the cell starting from \code{1}.}

\item{savePath}{A character string naming the absolute path of the directory
where the cell life images will be saved (excluding the last \code{"/"}).
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
}
\description{
Creates the life images of a set of given cells.
}
\details{
Separate life image(s) for each cell specified in \code{cells} are generated.
The cell is viewed in its colony for every frame of its lifespan and is marked as red.
The rest cells of the colony are marked as white.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} users, importing the data with \code{\link{import_basca}}.
\cr\cr
Users of \emph{Oufti} or \emph{SuperSegger} 
who imported the data with \code{\link{import_oufti}} or \code{\link{import_ss}}, respectively,
are \bold{excluded} from using this function, as no colony list was returned.
\cr\cr
If \code{\link{import_json}} was used for importing the data,
it is necessary that cell list elements have the \code{pixelList} and \code{colId} components
and colony list elements have the \code{colImage} component.
See \code{\link{import_json}} for more details.
In other case, this function cannot be used (throws an error).
}

