% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movie.R
\name{create_movie}
\alias{create_movie}
\title{Create movie}
\usage{
create_movie(overallLT, LT, cell_list, col_list, frameH, frameW, Ncols,
  attrC = "", unitC = "", NC = NULL, savePath = getwd())
}
\arguments{
\item{overallLT}{The lineage tree the cells of which will be visualized in the created movie,
an object of class \code{"igraph"}.
Cells that do not belong to the \code{LT} are colored white.
\cr\cr
For visualizing all cells extracted from the analysis of the movie,
use the tree returned from \code{\link{unite_trees}}
when called for all parts of the overall FLT (main part and possible motherless branches).}

\item{LT}{The lineage tree the cells of which will be colored in the created movie,
an object of class \code{"igraph"}.
This tree must be a subtree of the \code{overallLT} regarding the included cells in the analysis.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{frameH}{Frame image height in \emph{pixels}, a non-zero positive integer value.}

\item{frameW}{Frame image width in \emph{pixels}, a non-zero positive integer value.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{attrC}{The name of the attribute in the \code{LT} by which the cells will be colored, a character string.
It can be any numeric or boolean attribute, as returned from \code{\link{get_attr_names}}.
Coloring is applied to all non-root cells of the \code{LT}, as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in this attribute which are colored gray.
When the default value \code{""} (the empty character) is used, all cells of the \code{LT} are colored gray.
Cells that belong to the \code{overallLT} but not to the \code{LT} are always colored white.}

\item{unitC}{The unit of \code{attrC}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attrC} is in arbitrary units.
This argument is ignored in case \code{attrC = ""}.}

\item{NC}{Number of colonies in the movie (if \code{attrC = "colony"}) or
number of generations in the movie (if \code{attrC = "generation"}), a non-zero positive integer value.
This argument is ignored in case \code{attrC != "colony"}
and \code{attrC != "generation"}.}

\item{savePath}{A character string naming the absolute path of the directory
where the generated \code{.png} files will be saved.
The default value is the current working directory \code{getwd()}.
Image files are named as \code{"frame_<i>.png"}, where \code{"<i>"} is the frame ID.
The movie file is named as \code{"movie.mp4"}.
All files will be saved in folder \code{"<attrC>"}
(or \code{"noColor"} in case \code{attrC = ""}), created under the specified directory.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
}
\description{
Creates the movie of the segmented cells.
}
\details{
A separate image for each frame existing in the \code{overallLT} is generated.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} or \emph{SuperSegger} users,
importing the data with \code{\link{import_basca}} or \code{\link{import_ss}}, respectively.
\cr\cr
If \code{\link{import_json}} was used for importing the data,
it is necessary that cell list elements have the \code{pixelList} and \code{colId} components
and colony list elements have the \code{ULcorner} component.
See \code{\link{import_json}} for more details.
In other case, this function cannot be used (throws an error).
\cr\cr
Users of \emph{Oufti} who imported the data with \code{\link{import_oufti}}
are also \bold{excluded} from using this function, as no colony list was returned.

\cr\cr
\emph{FFmpeg} is \bold{required} in the system in order to automatically
convert the series of the generated \code{.png} files to an \code{.mp4} file,
at 5 fps (frames per second).
Image files for more than two consequtive frames starting from frame \code{1}
must have been generated.
}

\seealso{
\code{\link{isSubtree}} for checking if a tree is a subtree of another tree.
}
