% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFDT.R
\name{createFDT}
\alias{createFDT}
\title{Create cells' FDT}
\usage{
createFDT(LTmain, minLife = 5, frameR)
}
\arguments{
\item{LTmain}{The main part of the overall FLT,
a connected lineage tree containing the imaginary \emph{root} cells
(object of class \code{"igraph"}).}

\item{minLife}{Minimum life in \emph{frames} for a cell to be considered in the analysis, a positive integer value.
The default value is \code{5}. Use value \code{0} to exclude all leaf cells from the analysis.}

\item{frameR}{Frame rate of the movie in \emph{frames} per \emph{minute}, a non-zero positive numeric value.}
}
\value{
A named list with the following components:
\item{DTmain}{The corresponding main part of the overall FDT,
a connected division tree containing the imaginary \emph{root} cells (object of class \code{"igraph"}).}
\item{LTmain}{The updated LTmain with the attributes \code{"generation"} and \code{"age"} added/updated,
an object of class \code{"igraph"}.}
\item{Ngens}{Number of generations in the movie, a non-zero positive integer value.
IDs of generations are in the range \code{[0, Ngens-1]}.}
}
\description{
Creates the cells' Forest of Division Trees (FDT) given the corresponding Forest of Lineage Trees (FLT).
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected,
\code{\link{save_tree}} for saving a tree on disc.
}
