% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_attr.R
\name{del_attr}
\alias{del_attr}
\title{Delete an attribute}
\usage{
del_attr(tree, attr)
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
It can be any numeric, boolean or character string attribute, as returned from \code{\link{get_attr_names}}.}
}
\value{
The updated tree with the attribute \code{attr} deleted, an object of class \code{"igraph"}.
}
\description{
Deletes an attribute from a lineage or division tree.
}
