% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_branch.R
\name{extract_branch}
\alias{extract_branch}
\title{Extract a branch from a tree}
\usage{
extract_branch(tree, cell)
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{cell}{The label of the cell in the \code{tree} which is the root of the branch to be extracted, a character string.
It can be any cell, as returned from \code{\link{get_cells}}.}
}
\value{
A named list with the following components:
\item{treeNew}{The new tree with the branch extracted, an object of class \code{"igraph"}.}
\item{branch}{The extracted connected motherless branch, an object of class \code{"igraph"}.}
}
\description{
Extracts a branch from a lineage or division tree. The branch for extraction is defined by its root cell.
}
\seealso{
\code{\link{add_branch}} for connecting a motherless branch to a lineage tree.
}
