% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr_names.R
\name{get_attr_names}
\alias{get_attr_names}
\title{Get attributes' names}
\usage{
get_attr_names(tree, type = c("n", "b", "c"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{type}{A character string naming the type of attributes to be returned:
\itemize{
\item \code{"n"} for numeric attributes
\item \code{"b"} for boolean attributes
\item \code{"c"} for character string attributes
}}
}
\value{
The correponding attributes' names, a vector of character strings.
}
\description{
Returns the names of the attributes of a lineage or division tree.
}
