% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr_stats.R
\name{get_attr_stats}
\alias{get_attr_stats}
\title{Get statistic of an attribute}
\usage{
get_attr_stats(tree, treeT = c("LT", "DT"), attr, stat = c("mean", "median",
  "sd", "min", "max"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.}

\item{stat}{A character string naming the statistic to be returned:
\itemize{
\item \code{"mean"} for the \emph{mean}
\item \code{"median"} for the \emph{median}
\item \code{"sd"} for the \emph{standard deviation}
\item \code{"min"} for the minimum value
\item \code{"max"} for the maximum value
}}
}
\value{
A named list with the following components:
\item{Ncells}{Number of cells, a positive integer value.}
\item{value}{The corresponding statistic of attribute \code{attr}, a numeric value,
or \code{NA}.}
}
\description{
Returns the statistic of a numeric attribute of a lineage or division tree.
}
\details{
The statistic is calculated considering all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
}
