% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr_vals.R
\name{get_attr_vals}
\alias{get_attr_vals}
\title{Get attribute's values}
\usage{
get_attr_vals(tree, attr)
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
It can be any numeric, boolean or character string attribute, as returned from \code{\link{get_attr_names}}.}
}
\value{
A vector with the values of attribute \code{attr}.
}
\description{
Returns the values of an attribute for all cells of a lineage or division tree,
as returned from \code{\link{get_cells}}.
}
