% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cand_merge_cells.R
\name{get_cand_merge_cells}
\alias{get_cand_merge_cells}
\title{Find candidate merge cells}
\usage{
get_cand_merge_cells(LT, cell, LTcand = NULL, cell_list, col_list, Ncols,
  show = TRUE)
}
\arguments{
\item{LT}{The lineage tree where the cell specified in \code{cell} belongs, an object of class \code{"igraph"}.}

\item{cell}{The label of the cell in the \code{LT} for which to find candidate merge cells, a character string.
It can be any non-root cell, as returned from \code{\link{get_cells}}.}

\item{LTcand}{The lineage tree where the candidate merge cells of \code{cell} will belong, an object of class \code{"igraph"}.
When the default value \code{NULL} is used, candidate merged cells will belong in the \code{LT}.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{show}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether \code{\link{view_cell}} will be called for
the \code{cell} and the candidate merge cells.
This capability is useful in order to choose the desired merge cell and call \code{\link{merge_cells}}.
The default value is \code{TRUE}.}
}
\value{
The labels of the candidate merge cells, a vector of character strings.
In case no candidate merge cells are found, \code{NULL} is returned.
}
\description{
Returns the candidate cells for a given cell to be merged with, in the same or different lineage tree.
}
\details{
Candidate merge cells are cells of the same colony and frame with least one common pixel with the given cell.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} users \bold{only},
importing the data with \code{\link{import_basca}}.
}

