% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cand_mother_cells.R
\name{get_cand_mother_cells}
\alias{get_cand_mother_cells}
\title{Find candidate mothers}
\usage{
get_cand_mother_cells(branch, LT, Nd = 1, Ncands = 5, cell_list, col_list,
  Ncols, show = FALSE)
}
\arguments{
\item{branch}{The connected lineage tree (motherless branch) for which to find candidate mothers, an object of class \code{"igraph"}.}

\item{LT}{The lineage tree where the candidate mothers of the \code{branch} will belong, an object of class \code{"igraph"}.}

\item{Nd}{Maximum number of daughters for a cell in the \code{LT} to be considered as a candidate mother, a positive integer value.
The default value is \code{1}.}

\item{Ncands}{Maximum number of candidate mothers to be returned, a non-zero positive integer value.
The default value is \code{5}. Use value \code{NULL} to return all candidate mothers.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{show}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether \code{\link{view_cell}}
will be called for the root of the \code{branch} and the candidate mothers.
This capability is useful in order to choose the desired mother cell and call \code{\link{add_branch}}.
The default value is \code{FALSE}.
See the \emph{Prerequisites} section of \code{\link{view_cell}} in order to be able to use value \code{TRUE}.}
}
\value{
The labels of the candidate mothers, a vector of character strings.
Candidate mothers are sorted by ascending distance between their centroid and
the centroid of the root of the \code{branch}, which implies best matching.
In case no candidate mothers are found, \code{NULL} is returned.
}
\description{
Returns the candidate mother cells for the root of a motherless branch in a lineage tree.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} or \emph{SuperSegger} users,
importing the data with \code{\link{import_basca}} or \code{\link{import_ss}}, respectively.
\cr\cr
If \code{\link{import_json}} was used for importing the data,
it is necessary that cell list elements have the \code{centroid} and \code{colId} components
and colony list elements have the \code{ULcorner} component.
See \code{\link{import_json}} for more details.
In other case, this function cannot be used (throws an error).
\cr\cr
Users of \emph{Oufti} who imported the data with \code{\link{import_oufti}}
are also \bold{excluded} from using this function, as no colony list was returned.

Candidate mothers are cells of the same colony in the previous frame.
}

\seealso{
\code{\link{isConnected}} for checking if a tree is connected.
}
