% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_attr_vals.R
\name{get_cell_attr_vals}
\alias{get_cell_attr_vals}
\title{Get attributes' values of a cell}
\usage{
get_cell_attr_vals(tree, treeT = c("LT", "DT"), cell)
}
\arguments{
\item{tree}{The lineage or division tree where the cell specified in \code{cell} belongs,
an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{cell}{The label of the cell in the \code{tree} whose attributes' values will be returned, a character string.
It can be any non-root cell, as returned from \code{\link{get_cells}}.}
}
\value{
The attributes' values of the \code{cell}, a named list.
}
\description{
Returns the values of the attributes of a cell in a lineage or division tree.
}
