% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_fam.R
\name{get_cell_fam}
\alias{get_cell_fam}
\title{Get family cells of a cell}
\usage{
get_cell_fam(tree, treeT = c("LT", "DT"), cell, type = c("m", "gm", "d",
  "gd", "s", "c"))
}
\arguments{
\item{tree}{The connected lineage or division tree where the cell specified in \code{cell} belongs, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{cell}{The label of the cell in the \code{tree} whose family cells will be returned, a character string.
It can be any non-root cell, as returned from \code{\link{get_cells}}.}

\item{type}{A character string naming the type of the family cells to be returned:
\itemize{
\item \code{"m"} for mother
\item \code{"gm"} for grandmother
\item \code{"d"} for daughters
\item \code{"gd"} for granddaughters
\item \code{"s"} for sibling
\item \code{"c"} for cousins
}}
}
\value{
The labels of the correponding family cells of the \code{cell}, a vector of character strings.
In case no family cells are found, \code{NULL} is returned.
}
\description{
Returns the labels of the family cells of a cell in a lineage or division tree.
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected.
}
