% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cells.R
\name{get_cells}
\alias{get_cells}
\title{Get cells of a tree}
\usage{
get_cells(tree, treeT = c("LT", "DT"), type = c("all", "nr", "inc"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}
This argument is ignored in case \code{type = "all"}}

\item{type}{A character string naming the type of cells to be returned:
\itemize{
\item \code{"all"} for all cells (including any existing imaginary \emph{root} cells)
\item \code{"nr"} for all non-root cells (excluding any existing imaginary \emph{root} cells)
\item \code{"inc"} for all cells included in the analysis
}}
}
\value{
The labels of the corresponding cells, a vector of character strings.
}
\description{
Returns the labels of the cells in a lineage or division tree.
}
