% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LT_cell.R
\name{get_LT_cell}
\alias{get_LT_cell}
\title{Get the label of an instant of a cell}
\usage{
get_LT_cell(DT, cell, instant)
}
\arguments{
\item{DT}{The division tree where the cell specified in \code{cell} belongs,
an object of class \code{"igraph"}.}

\item{cell}{The label of the cell in the \code{DT}, a character string.
It can be any non-root cell, as returned from \code{\link{get_cells}}.}

\item{instant}{The instant of the cell specified in \code{cell},
a non-zero positive integer value.}
}
\value{
The label of the \code{cell} \code{instant} in the corresponding lineage tree of the \code{DT},
a character string.
In case \code{cell} has less instants than the specified \code{instant}, \code{NULL} is returned.
}
\description{
Returns the label of an instant of a cell in a lineage tree,
given the label of the cell in the corresponding division tree.
}
\seealso{
\code{\link{get_DT_cell}} for the reverse.
}
