% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConnected.R
\name{isConnected}
\alias{isConnected}
\title{Check if a tree is connected}
\usage{
isConnected(tree)
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}
}
\value{
A logical value (\code{TRUE} or \code{FALSE})
indicating whether the \code{tree} is connected or not.
}
\description{
Checks if a lineage or division tree is connected.
A tree is connected if only one motherless cell exists (i.e. the root of the tree).
}
