% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cells.R
\name{merge_cells}
\alias{merge_cells}
\title{Merge two cells}
\usage{
merge_cells(LT1, LT2 = NULL, cell1, cell2, cell_list, col_list, Ncols, pixelR,
  matFolder, matFileName, exeFolder, mcrFolder, show = TRUE)
}
\arguments{
\item{LT1}{The lineage tree where the cell specified in \code{cell1} belongs, an object of class \code{"igraph"}.}

\item{LT2}{The lineage tree where the cell specified in \code{cell12} belongs, an object of class \code{"igraph"}.
When the default value \code{NULL} is used, \code{cell2} belongs to the \code{LT}.}

\item{cell1}{The label of the first cell in the \code{LT1} to be merged, a character string.
It can be any non-root cell, as returned from \code{\link{get_cells}}.}

\item{cell2}{The label of the second cell in the \code{LT2} (or \code{LT1} in case \code{LT2 = NULL}) to be merged, a character string.
It can be any valid candidate merge cell, as returned from \code{\link{get_cand_merge_cells}}.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{pixelR}{The pixel ratio in units of length, a non-zero positive numeric value.}

\item{matFolder}{A character string naming the absolute path of the directory where the \code{.mat} file generated by \emph{BaSCA} is saved.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}

\item{matFileName}{A character string naming the \code{.mat} file generated by \emph{BaSCA} (including the suffix \code{".mat"}).
The filename is relative to the \code{matFolder}.}

\item{exeFolder}{A character string naming the absolute path of the installation folder of the MATLAB executable.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}

\item{mcrFolder}{A character string naming the absolute path of the installation folder of the Matlab Compiler Runtime (MCR).
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}

\item{show}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether \code{\link{view_cell}}
will be called for \code{cell1} and \code{cell2} before the merge operation
and the resulting cell after the merge operation.
This capability is useful in order to see the result of the function.
The default value is \code{TRUE}.}
}
\value{
A named list with the following components:
\item{LT1}{The updated LT1 with \code{cell1} replaced and the possible daughter branches added/replaced,
an object of class \code{"igraph"}.}
\item{LT2}{The updated LT2 with \code{cell2} and its daughter branches deleted,
an object of class \code{"igraph"}.
\code{NULL} is returned in case LT2 was anywise \code{NULL} or if it ended up with no cells.}
\item{cell_list}{The updated cell_list with \code{cell1} replaced.}
\item{branches}{A list with the remaining motherless branches.
Each branch (element of the list) is an object of class \code{"igraph"}.}
}
\description{
Merges two cells from the same or different lineage trees.
}
\details{
After the merge operation, \code{cell1} is replaced in both \code{LT1} and \code{cell_list} by the resulting cell.
Daughter branches of \code{cell2} are extracted from \code{LT2} (or \code{LT1} in case \code{LT2 = NULL}),
with successive calls of \code{\link{extract_branch}}.
These motherless branches (lineage trees) are added as daughter branches to \code{cell1},
with successive calls of \code{\link{add_branch}}, until \code{cell1} has two daughters.
Finally, \code{cell2} is deleted from \code{LT2} (or \code{LT1} in case \code{LT2 = NULL})
with \code{\link{extract_branch}}.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} users \bold{only},
importing the data with \code{\link{import_basca}}.
}

\seealso{
\code{\link{split_cell}} for the reverse.
}
