% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baranyi.R
\name{plot_baranyi}
\alias{plot_baranyi}
\title{Plot growth curves of cell counts}
\usage{
plot_baranyi(LT, DT, cols = -1, Ncols, Nframes, frameR, showRaw = FALSE,
  save = FALSE, savePars = list(w = 3000, h = 2000, res = 300, path =
  getwd(), name = "my_baranyi"))
}
\arguments{
\item{LT}{The lineage tree, an object of class \code{"igraph"}.}

\item{DT}{The corresponding division tree of the \code{LT},
an object of class \code{"igraph"}.}

\item{cols}{The IDs of the colonies for which to fit the \emph{Baranyi and Roberts} model,
a vector of non-zero positive integer values.
The default value \code{-1} stands for all existing colonies in the \code{LT}.
Use value \code{-2} for fitting the model to the whole population.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.
This argument is ignored in case \code{cols = -2}.}

\item{Nframes}{Number of frames in the movie, a non-zero positive integer value.}

\item{frameR}{Frame rate of the movie in \emph{frames} per \emph{minute}, a non-zero positive numeric value.}

\item{showRaw}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
the raw data (unfitted data points) will be shown on the plot or not, respectively.
The default value is \code{FALSE}.
This argument is ignored (regarded as \code{FALSE}) in case \code{cols = -1}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{3000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{300}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_baranyi"}.}
}}
}
\value{
A dataframe with the following columns:
\enumerate{
\item \code{colony} is the colony ID (a non-zero positive integer value)
or \code{-2} in case \code{cols = -2}.
\item \code{lag} is the \emph{\lambda} parameter (\emph{lag time})
of the \emph{Baranyi and Roberts} model in \emph{hours} (a non-zero positive numeric value)
or \code{NA} in case the model failed to be fitted or \code{colony} has only cells which have not been divided.
\item \code{mumax} is the \emph{mumax} parameter (\emph{maximum specific growth rate})
of the \emph{Baranyi and Roberts} model in \emph{1/hour} (a non-zero positive numeric value)
or \code{NA} in case the model failed to be fitted or \code{colony} has only cells which have not been divided.
}
For groups with \code{lag = NA}, no plot is generated, except for the case that \code{showRaw = TRUE}.
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Computes and plots the growth curves of cell counts (number of cells) in a lineage tree.
The growth curve can be plotted for each colony or for the whole population.
Each growth curve is computed by fitting the \emph{Baranyi and Roberts} model to the corresponding data.
}
\details{
The parameters of the \emph{Baranyi and Roberts} model are found using the
\emph{non-linear least squares} method provided by \code{\link[stats]{nls}},
considering all non-root cells, as returned from \code{\link{get_cells}}.
\cr\cr
When \code{cols = -1}, a common plot for all colonies is generated.
In other cases, a separate plot for each colony specified in \code{cols} is generated.
\cr\cr
x-axis represents the time in \emph{hours} from the start (value \code{0}) to end of the movie.
y-axis represents the cell counts in logarithmic scale.
The range of y-axis values depicted in each plot is common and
is calculated as the range of values of all specified colonies,
with minimum upper limit \code{2} (i.e. 20 cells).
\cr\cr
Color denotes the corresponding colony.
}
