% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_col_tree.R
\name{plot_col_tree}
\alias{plot_col_tree}
\title{Plot colonies' lineage tree}
\usage{
plot_col_tree(col_list, Ncols, sizeV = 2, sizeE = 0.7, sizeL = 1,
  save = FALSE, savePars = list(w = 2000, h = 2000, res = 250, path =
  getwd(), name = "my_col_tree"))
}
\arguments{
\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}

\item{sizeV}{Size of vertices, a non-zero positive numeric value.
The default value is \code{2}.}

\item{sizeE}{Width of edges, a non-zero positive numeric value.
The default value is \code{0.7}.}

\item{sizeL}{Size of explanatory legends and title, a non-zero positive numeric value.
The default value is \code{1}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{250}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_col_tree"}.}
}}
}
\description{
Plots the colonies' lineage tree.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} or \emph{SuperSegger} users,
importing the data with \code{\link{import_basca}} or \code{\link{import_ss}}, respectively.
\cr\cr
Users of \emph{Oufti} who imported the data with \code{\link{import_oufti}}
are \bold{excluded} from using this function, as no colony list was returned.
\cr\cr
For other users, it is necessary that a colony list was imported with \code{\link{import_json}}.
In other case, no colony list exists and this function cannot be used.

Nodes represent the colony instants of the movie and are colored based on the ID of the corresponding colony.
The imaginary \emph{root} colony instant is colored white.
Gray nodes are colony instants which have arised from merged colonies.
}

