% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_attr2.R
\name{plot_dot_attr2}
\alias{plot_dot_attr2}
\title{Create scatter plot of two attributes}
\usage{
plot_dot_attr2(tree, treeT = c("LT", "DT"), attr1, unit1 = "", attr2,
  unit2 = "", attrC = "", unitC = "", NC = NULL, attrS = "",
  Nbins = c(20, 20), save = FALSE, savePars = list(w = 2500, h = 2000, res
  = 350, path = getwd(), name = "my_dot_attr2"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr1, attr2}{The names of the attributes in the \code{tree}, character strings.
Each one can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.}

\item{unit1, unit2}{The units of the corresponding \code{attr}, character strings.
Each one should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value of each is the empty character \code{""},
which implies that the corresponding \code{attr} is in arbitrary units.}

\item{attrC}{The name of the attribute in the \code{tree} by which the cells will be colored, a character string.
It can be any numeric or boolean attribute, as returned from \code{\link{get_attr_names}}.
Coloring is applied to all depicted cells,
except for cells with \code{NA} value in this attribute which are colored gray.
When the default value \code{""} (the empty character) is used, data points' colors are the default.}

\item{unitC}{The unit of \code{attrC}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attrC} is in arbitrary units.
This argument is ignored in case \code{attrC = ""}.}

\item{NC}{Number of colonies in the movie (if \code{attrC = "colony"}) or
number of generations in the movie (if \code{attrC = "generation"}), a non-zero positive integer value.
This argument is ignored in case \code{attrC != "colony"}
and \code{attrC != "generation"}.}

\item{attrS}{The name of the attribute in the \code{tree} by which the cells will be shaped, a character string.
It can be any boolean attribute, as returned from \code{\link{get_attr_names}}.
Cells are represented by dots or squares if their value in this attribute is \code{TRUE} or \code{FALSE}, respectively.
When the default value \code{""} (the empty character) is used, data points' shape is the default.}

\item{Nbins}{Number of equally spaced bins to be used for \code{attr1} and \code{attr2},
a vector of two non-zero positive integer values \code{>=2}.
The first value is used for \code{attr1} and the second for \code{attr2}.
These values are indicative and may change automatically depending on the values of \code{attr1} and \code{attr2},
producing a warning message.
The default value is \code{c(20, 20)}.
This argument is ignored in case \code{attrC != ""}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{350}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_dot_attr2"}.}
}}
}
\value{
A named list with the following components:
\item{Ncells}{Number of cells, a non-zero positive integer value.}
\item{r}{The Pearson correlation coefficient (a numeric value in the range \code{[-1, 1]})
or \code{NA} in case less than 2 unique data points exist.}
\item{regression}{A named list with the following components:
\itemize{
\item \code{a} is the slope of the regression line, a numeric value
\item \code{b} is the y-intercept of the regression line, a numeric value
\item \code{r2} is the R-squared coefficient of the regression line
as returned from \code{\link[stats]{summary.lm}}, a numeric value in the range \code{[0, 1]}
}
In case less than 2 unique data points exist, \code{NULL} is returned, instead.}
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Creates the X-Y scatter plot of two numeric attributes of a lineage or division tree.
}
\details{
Each data point (x,y) represents the value of \code{attr1} and \code{attr2} of a cell, respectively.
The scatter plot is created for all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr1} or \code{attr2}.
\cr\cr
By default, all data points are dot points colored based on the density of the (X,Y) variable value.
2D binning is applied on the range of all \code{attr1} and \code{attr2} values.
Density represents the counts (number of data points) in each bin.
\cr\cr
The linear regression line is also drawn on the plot.
X is the predictor variable and Y is the response.
The parameters of the regression line are found using the \emph{linear least squares} method
provided by \code{\link[stats]{lm}}.
}
