% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_attr2_gen2.R
\name{plot_dot_attr2_gen2}
\alias{plot_dot_attr2_gen2}
\title{Create scatter plot of attribute(s) between two generations}
\usage{
plot_dot_attr2_gen2(DT, gen1, gen2, attr1, unit1 = "", attr2 = "",
  unit2 = "", save = FALSE, savePars = list(w = 2500, h = 2000, res = 350,
  path = getwd(), name = "my_dot_attr2_gen2"))
}
\arguments{
\item{DT}{The connected division tree, an object of class \code{"igraph"}.}

\item{gen1}{The ID of the first generation in the \code{DT}, a positive integer value.}

\item{gen2}{The ID of the second generation in the \code{DT}, a positive integer value.}

\item{attr1}{The name of the first attribute in the \code{DT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"} and \code{"generation"}.
When \code{attr2 = ""}, this attribute is used for cells of both generations \code{gen1} and \code{gen2}.
In other case, this attribute is used for cells of generation \code{gen1}.}

\item{unit1}{The unit of \code{attr1}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr1} is in arbitrary units.}

\item{attr2}{The name of the second attribute in the \code{DT}, a character string.
When the default value \code{""} (the empty string) is used,
the same attribute \code{attr1} is used for cells of both generations \code{gen1} and \code{gen2}.
In other case, it can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"} and \code{"generation"}, and it is used for cells of generation \code{gen2}.}

\item{unit2}{The unit of \code{attr2}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr2} is in arbitrary units.
This argument is ignored in case \code{attr2 = ""}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{350}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_dot_attr2_gen2"}.}
}}
}
\value{
A named list with the following components:
\item{Ncells}{Number of cells, a non-zero positive integer value.}
\item{r}{The Pearson correlation coefficient (a numeric value in the range \code{[-1, 1]})
or \code{NA} in case less than 2 unique data points exist.}
\item{regression}{A named list with the following components:
\itemize{
\item \code{a} is the slope of the regression line, a numeric value
\item \code{b} is the y-intercept of the regression line, a numeric value
\item \code{r2} is the R-squared coefficient of the regression line
as returned from \code{\link[stats]{summary.lm}}, a numeric value in the range \code{[0, 1]}
}
In case less than 2 unique data points exist, \code{NULL} is returned, instead.}
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Creates the X-Y scatter plot of the same or different numeric attribute
between cells of two specific generations of a division tree.
}
\details{
Each data point (x,y) represents the corresponding attribute value of
the older and younger cell, respectively.
The scatter plot is created for all specified cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr1} (or \code{attr2} in case \code{attr2 != ""}).
\cr\cr
Data points are purple dot points.
\cr\cr
The linear regression line is also drawn on the plot.
X is the predictor variable and Y is the response.
The parameters of the regression line are found using the \emph{linear least squares} method
provided by \code{\link[stats]{lm}}.
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected.
}
