% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_attr_fam.R
\name{plot_dot_attr_fam}
\alias{plot_dot_attr_fam}
\title{Create scatter plot of an attribute between family cells}
\usage{
plot_dot_attr_fam(DT, attr, unit = "", type = c("s", "c", "md", "gmgd"),
  gens = -1, Ngens, save = FALSE, savePars = list(w = 2500, h = 2000, res
  = 350, path = getwd(), name = "my_dot_attr_fam"))
}
\arguments{
\item{DT}{The connected division tree, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{DT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"} and \code{"generation"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{type}{A character string naming the type of the family cells for which to create the scatter plot:
\itemize{
\item \code{"s"} for siblings
\item \code{"c"} for cousins
\item \code{"md"} for mother and daughter
\item \code{"gmgd"} for grand-mother and grand-daughter
}}

\item{gens}{The IDs of the generations which will be included in the scatter plot,
a vector of non-zero positive integer values.
When \code{type = "md"} or \code{type = "gmgd"},
each value denotes the generation of the younger cell of the family (i.e. daughter or grand-daughter, respectively).
Acceptable values are in the range \code{[1, Ngens-1]} in case \code{type = "s"} or \code{type = "md"}
and in the range \code{[2, Ngens-1]} in case \code{type = "c"} or \code{type = "gmgd"}.
The default value \code{-1} stands for all existing generations in the \code{DT} (whole population).}

\item{Ngens}{Number of generations in the movie, a non-zero positive integer value.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{350}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_dot_attr_fam"}.}
}}
}
\value{
A named list with the following components:
\item{Ncells}{Number of cells, a non-zero positive integer value.}
\item{r}{The Pearson correlation coefficient (a numeric value in the range \code{[-1, 1]})
or \code{NA} in case less than 2 unique data points exist.}
\item{regression}{A named list with the following components:
\itemize{
\item \code{a} is the slope of the regression line, a numeric value
\item \code{b} is the y-intercept of the regression line, a numeric value
\item \code{r2} is the R-squared coefficient of the regression line
as returned from \code{\link[stats]{summary.lm}}, a numeric value in the range \code{[0, 1]}
}
In case less than 2 unique data points exist, \code{NULL} is returned, instead.}
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Creates the X-Y scatter plot of a numeric attribute between family cells of a division tree.
The scatter plot can be created for specific generation(s) or for the whole population.
}
\details{
A common scatter plot for all generations specified in \code{gens} is created.
In the general case, each data point (x,y) represents the attribute value of
the older and younger cell of the family, respectively.
The scatter plot is created for all specified family cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
\cr\cr
Data points are dot points colored based on the generation of the younger cell.
Note that when \code{type = "s"} or \code{type = "c"}, there is no distinction between older and younger cell,
since both cells are of the same generation.
\cr\cr
The linear regression line is also drawn on the plot.
X is the predictor variable and Y is the response.
The parameters of the regression line are found using the \emph{linear least squares} method
provided by \code{\link[stats]{lm}}.
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected.
}
