% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot_time_attr.R
\name{plot_dot_time_attr}
\alias{plot_dot_time_attr}
\title{Create dot plot of an attribute by time}
\usage{
plot_dot_time_attr(tree, treeT = c("LT", "DT"), attr, unit = "",
  Nbins = 20, save = FALSE, savePars = list(w = 2500, h = 2000, res = 350,
  path = getwd(), name = "my_dot_time_attr"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
In case \code{treeT = "LT"}, it can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.
In case \code{treeT = "DT"}, it can be any numeric attribute, as returned from \code{\link{get_attr_names}},
having the suffix \code{"_birth"} or \code{"_division"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{Nbins}{Number of equally spaced bins to be used for \code{attr}, a non-zero positive integer value \code{>=2}.
This value is indicative and may change automatically depending on the values of \code{attr},
producing a warning message.
The default value is \code{20}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{350}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_dot_time_attr"}.}
}}
}
\value{
A named list with the following components:
\item{Ncells}{Number of cells, a non-zero positive integer value.}
\item{r}{The Pearson correlation coefficient (a numeric value in the range \code{[-1, 1]})
or \code{NA} in case less than 2 unique data points exist.}
\item{regression}{A named list with the following components:
\itemize{
\item \code{a} is the slope of the regression line, a numeric value
\item \code{b} is the y-intercept of the regression line, a numeric value
\item \code{r2} is the R-squared coefficient of the regression line
as returned from \code{\link[stats]{summary.lm}}, a numeric value in the range \code{[0, 1]}
}
In case less than 2 unique data points exist, \code{NULL} is returned, instead.}
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Creates the dot plot of a numeric attribute of a lineage or division tree by time.
}
\details{
Each data point (x,y) represents the time in \emph{frames} and the value of \code{attr} of a cell, respectively.
The dot plot is created for all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
\cr\cr
Data points are dot points colored based on the density of the Y variable value at every time point x.
Binning is applied on the range of all \code{attr} values.
Density represents the counts (number of data points) in each bin at every time point x.
Ultimately, the plot is a color representation of the histogram of \code{attr} in time.
}
