% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_attr.R
\name{plot_growth_attr}
\alias{plot_growth_attr}
\title{Plot raw growth curves of an attribute}
\usage{
plot_growth_attr(DT, LT, attr, unit = "", cells = "all", Nframes,
  save = FALSE, savePars = list(w = 2500, h = 2000, res = 300, path =
  getwd(), name = "my_growth_attr"))
}
\arguments{
\item{DT}{The division tree where the cells specified in \code{cells} belong, an object of class \code{"igraph"}.}

\item{LT}{The corresponding lineage tree of the \code{DT}, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{LT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"}, \code{"frame"} and \code{"age"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{cells}{The labels of the cells in the \code{DT} whose growth curve will be plotted,
a vector of character strings.
They can be any cells that are included in the analysis, as returned from \code{\link{get_cells}}.
The default value \code{"all"} stands for all cells in the \code{DT}.}

\item{Nframes}{Number of frames in the movie, a non-zero positive integer value.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{300}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_growth_attr"}.}
}}
}
\description{
Plots the raw single-cell growth curves (unfitted time-series data) of a numeric attribute
for a set of given cells in a division tree.
}
\details{
A common plot for all cells specified in \code{cells} is generated.
\cr\cr
x-axis represents the time in \emph{frames},
from the start (value \code{1}) to end of the movie (value \code{Nframes}).
\cr\cr
Each single-cell growth curve is randomly colored.
}
