% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_attr_cell.R
\name{plot_growth_attr_cell}
\alias{plot_growth_attr_cell}
\title{Plot single-cell growth curve of an attribute}
\usage{
plot_growth_attr_cell(DT, LT, attr, unit = "", cells = "all",
  model = c("", "lin", "exp", "both"), frameR, save = FALSE,
  savePars = list(w = 1500, h = 1000, res = 150, path = getwd(), name =
  "my_growth_attr_cell"))
}
\arguments{
\item{DT}{The division tree where the cells specified in \code{cells} belong, an object of class \code{"igraph"}.}

\item{LT}{The corresponding lineage tree of the \code{DT}, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{LT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"}, \code{"frame"} and \code{"age"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{cells}{The labels of the cells in the \code{DT} whose growth curve will be plotted,
a vector of character strings.
They can be any non-root cells, as returned from \code{\link{get_cells}}.
The default value \code{"all"} stands for all cells in the \code{DT}.}

\item{model}{A character string naming the optional fitted growth curve(s) to be plotted,
if the corresponding parameters have been already successfully estimated
using \code{\link{add_attr_growth_fit_pars}}:
\itemize{
\item \code{""} (the empty string) for plotting just the raw growth curve
\item \code{"lin"} for plotting the fitted \emph{linear} model
\item \code{"exp"} for plotting the fitted \emph{exponential} model
\item \code{"both"} for plotting both fitted \emph{linear} and \emph{exponential} model
}}

\item{frameR}{Frame rate of the movie in \emph{frames} per \emph{minute}, a non-zero positive numeric value.
This argument is ignored in case \code{model = ""}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{1500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{1000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{150}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_growth_attr_cell"}.}
}}
}
\description{
Plots the raw (and fitted) single-cell growth curve (time-series data) of a numeric attribute
for a set of given cells in a division tree.
}
\details{
A separate plot for each cell specified in \code{cells} is generated.
\cr\cr
x-axis represents the life of the cell in \emph{frames},
starting from its birth (value \code{1}).
}
