% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_growth_attr_fit.R
\name{plot_growth_attr_fit}
\alias{plot_growth_attr_fit}
\title{Plot fitted growth curves of an attribute}
\usage{
plot_growth_attr_fit(DT, LT, attr, unit = "", model = c("lin", "exp"),
  grouped = c("col", "gen", "pop"), groups = -1, Ngroups, attrC = "",
  unitC = "", NC = NULL, dur = 1, sizeL = 0.7, save = FALSE,
  savePars = list(w = 1500, h = 1000, res = 150, path = getwd(), name =
  "my_growth_attr_fit"))
}
\arguments{
\item{DT}{The division tree, an object of class \code{"igraph"}.}

\item{LT}{The corresponding lineage tree of the \code{DT}, an object of class \code{"igraph"}.}

\item{attr}{The name of the attribute in the \code{LT}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"}, \code{"frame"} and \code{"age"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units..}

\item{model}{A character string naming the type of the fitted growth models to be plotted,
\itemize{
\item \code{"lin"} for plotting the fitted \emph{linear} models
\item \code{"exp"} for plotting the fitted \emph{exponential} models
}
Parameters of the model must have already been estimated using \code{\link{add_attr_growth_fit_pars}}.}

\item{grouped}{A character string naming the grouping method:
\itemize{
\item \code{"col"} for grouping by colony
\item \code{"gen"} for grouping by generation
\item \code{"pop"} for no grouping (whole population)
}}

\item{groups}{The IDs of the groups for which to create the plot, a vector of positive integer values.
This argument is ignored in case \code{grouped = "pop"}.
The default value \code{-1} stands for all existing groups in the \code{DT}.}

\item{Ngroups}{Number of colonies in the movie (if \code{grouped = "col"}) or
number of generations in the movie (if \code{grouped = "gen"}), a non-zero positive integer value.
This argument is ignored in case \code{grouped = "pop"}.}

\item{attrC}{The name of the attribute in the \code{DT} by which the cells' growth curves will be colored,
a character string.
It can be any numeric or boolean attribute, as returned from \code{\link{get_attr_names}}.
Coloring is applied to all depicted cells of all groups specified in \code{groups},
except for cells with \code{NA} value in this attribute which are colored gray.
When the default value \code{""} (the empty character) is used, cells' colors are the default.}

\item{unitC}{The unit of \code{attrC}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attrC} is in arbitrary units.
This argument is ignored in case \code{attrC = ""}.}

\item{NC}{Number of colonies in the movie (if \code{attrC = "colony"}) or
number of generations in the movie (if \code{attrC = "generation"}), a non-zero positive integer value.
This argument is ignored in case \code{attrC != "colony"}
and \code{attrC != "generation"}.}

\item{dur}{The time span in \code{hours}, a non-zero positive numeric value.
The default value is \code{1}.}

\item{sizeL}{Size of explanatory legends, a non-zero positive numeric value.
The default value is \code{0.7}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{1500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{1000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{150}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_growth_attr_fit"}.}
}}
}
\value{
A dataframe with the following columns:
\itemize{
\item When \code{model = "lin"}:
\enumerate{
\item \code{group} is the ID of the group (a positive integer value)
or \code{-2} in case \code{grouped = "pop"}
\item \code{Ncells} is the number of cells, a positive integer value
\item \code{a_mean} is the \emph{mean} of \code{"<attr>_a"}
(a non-zero positive numeric value in units of \code{attr} per \emph{hour}),
or \code{NA} in case \code{Ncells = 0}
\item \code{a_sd} is the \emph{standard deviation} of \code{"<attr>_a"}
(a non-zero positive numeric value in units of \code{attr} per \emph{hour}),
or \code{NA} in case \code{Ncells = 0} or \code{Ncells = 1}
\item \code{b_mean} is the \emph{mean} of \code{"<attr>_b"}
(a positive numeric value in units of \code{attr})
or \code{NA} in case \code{Ncells = 0}
\item \code{b_sd} is the \emph{standard deviation} of \code{"<attr>_b"}
(a positive numeric value in units of \code{attr})
or \code{NA} in case \code{Ncells = 0} or \code{Ncells = 1}
}
\item When \code{model = "exp"}:
\enumerate{
\item \code{group} is the ID of the group (a positive integer value)
or \code{-2} in case \code{grouped = "pop"}
\item \code{Ncells} is the number of cells, a positive integer value
\item \code{k_mean} is the \emph{mean} of \code{"<attr>_k"}
(a non-zero positive numeric value in units of \code{attr} per \emph{hour}),
or \code{NA} in case \code{Ncells = 0}
\item \code{k_sd} is the \emph{standard deviation} of \code{"<attr>_k"}
(a non-zero positive numeric value in units of \code{attr} per \emph{hour}),
or \code{NA} in case \code{Ncells = 0} or \code{Ncells = 1}
\item \code{y0_mean} is the \emph{mean} of \code{"<attr>_0"}
(a non-zero positive numeric value in units of \code{attr})
or \code{NA} in case \code{Ncells = 0}
\item \code{y0_sd} is the \emph{standard deviation} of \code{"<attr>_0"}
(a non-zero positive numeric value in units of \code{attr})
or \code{NA} in case \code{Ncells = 0} or \code{Ncells = 1}
}
}
\cr\cr
For groups with \code{Ncells = 0}, no plot of first or second type is generated.
For groups with \code{Ncells = 1}, no plot of second type is generated.
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Plots the fitted single-cell growth curves (fitted time-series data) of a numeric attribute
for all cells in a division tree.
The growth curves can be plotted per colony or generation or for the whole population.
The average growth curve of each group is also computed and plotted separately.
}
\details{
For each group specified in \code{groups}, two types of plots are generated.
The first plot depicts the single-cell growth curves of all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} values in the attributes/parameters of the \code{model},
as estimated by \code{\link{add_attr_growth_fit_pars}}.
By default, each single-cell growth curve is randomly colored.
\cr\cr
The second plot depicts the average growth curve of the group, +/- one standard deviation.
The following curves are drawn:
\itemize{
\item When \code{model = "lin"}:
\itemize{
\item \code{y = a_mean * t + b_mean}
\item \code{y = (a_mean + a_sd) * t + (b_mean + b_sd)}
\item \code{y = (a_mean - a_sd) * t + (b_mean - b_sd)}
}
\item When \code{model = "exp"}:
\itemize{
\item \code{y = y0_mean * e^(k_mean * t)}
\item \code{y = (y0_mean + y0_sd) * e^((k_mean + k_sd) * t)}
\item \code{y = (y0_mean - y0_sd) * e^((k_mean - k_sd) * t)}
}
}
The parameters of these curves are computed based on the corresponding parameters of the single-cell growth curves.
See the \emph{Value} field for more details.
Color of the area between the curves denotes the corresponding group.
\cr\cr
In both types of plots, x-axis represents the time in \emph{hours} in the range \code{[0, dur]}.
Value \code{0} is considered to be the birth time of each cell.
The range of y-axis values is common among the plots of the same type and
is calculated as the range of the corresponding values of all specified groups (excluding the outliers).
}
