% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_attr.R
\name{plot_hist_attr}
\alias{plot_hist_attr}
\title{Create histogram of an attribute}
\usage{
plot_hist_attr(tree, treeT = c("LT", "DT"), attr, unit = "",
  grouped = c("col", "gen", "pop"), groups = -1, Ngroups, Nbins = 20,
  save = FALSE, savePars = list(w = 2000, h = 2000, res = 250, path =
  getwd(), name = "my_hist_attr"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{grouped}{A character string naming the grouping method:
\itemize{
\item \code{"col"} for grouping by colony
\item \code{"gen"} for grouping by generation
\item \code{"pop"} for no grouping (whole population)
}}

\item{groups}{The IDs of the groups for which to create the histogram, a vector of positive integer values.
This argument is ignored in case \code{grouped = "pop"}.
The default value \code{-1} stands for all existing groups in the \code{tree}.}

\item{Ngroups}{Number of colonies in the movie (if \code{grouped = "col"}) or
number of generations in the movie (if \code{grouped = "gen"}), a non-zero positive integer value.
This argument is ignored in case \code{grouped = "pop"}.}

\item{Nbins}{Number of equally spaced bins to be used for \code{attr}, a non-zero positive integer value \code{>=2}.
This value is indicative and may change automatically depending on the values of \code{attr},
producing a warning message.
The default value is \code{20}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{250}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_hist_attr"}.}
}}
}
\value{
A dataframe with the following columns:
\enumerate{
\item \code{group} is the ID of the group (a positive integer value)
or \code{-2} in case \code{grouped = "pop"}.
\item \code{Ncells} is the number of cells, a positive integer value.
}
For groups with \code{Ncells = 0}, no plot is generated.
In case less than 2 unique cells exist, no plot is generated.
In case no cells exist, \code{NULL} is returned.
}
\description{
Creates the histogram of a numeric attribute of a lineage or division tree.
The histogram can be created for each colony or generation or for the whole population.
}
\details{
A separate plot for each group specified in \code{groups} is generated.
Each histogram is created considering all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
\cr\cr
y-axis represents the cell counts.
The range of x-axis values depicted in each plot is common and
is calculated as the range of \code{attr} values of all specified groups.
Binning is applied on the same common range.
\cr\cr
Color denotes the corresponding group.
}
