% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ncells.R
\name{plot_Ncells}
\alias{plot_Ncells}
\title{Plot cell counts}
\usage{
plot_Ncells(tree, treeT = c("LT", "DT"), grouped = c("col", "gen", "frame"),
  groups = -1, Ngroups, save = FALSE, savePars = list(w = 4000, h = 2000,
  res = 250, path = getwd(), name = "my_Ncells"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{grouped}{A character string naming the grouping method:
\itemize{
\item \code{"col"} for grouping by colony
\item \code{"gen"} for grouping by generation
\item \code{"frame"} for grouping by frame, which is acceptable in case \code{treeT = "LT"}.
}}

\item{groups}{The IDs of the groups for which to plot the cell counts, a vector of positive integer values.
The default value \code{-1} stands for all existing groups in the \code{tree}.}

\item{Ngroups}{Number of colonies in the movie (if \code{grouped = "col"}) or
number of generations in the movie (if \code{grouped = "gen"}) or
number of frames in the movie (if \code{grouped = "frame"}), a non-zero positive integer value.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{4000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{250}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved.
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_Ncells"}.}
}}
}
\value{
A dataframe with the following columns:
\enumerate{
\item \code{group} is the ID of the group, a positive integer value.
\item \code{Ncells} is the number of cells, a positive integer value.
}
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Calculates and plots the number of cells (cell counts) in a lineage or division tree.
Cell counts can be plotted for each colony, generation or frame.
}
\details{
A common plot for all groups specified in \code{groups} is generated.
Cell counts are calculated considering all cells that are included in the analysis,
as returned from \code{\link{get_cells}}.
\cr\cr
Color denotes the corresponding group.
}
