% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pdf_attr.R
\name{plot_pdf_attr}
\alias{plot_pdf_attr}
\title{Plot the PDF of an attribute}
\usage{
plot_pdf_attr(tree, treeT = c("LT", "DT"), attr, unit = "",
  grouped = c("col", "gen", "pop"), groups = -1, Ngroups,
  model = c("norm", "gamma", "lnorm", "auto"), plot3D = TRUE,
  save = FALSE, savePars = list(w = 2000, h = 2000, res = 250, path =
  getwd(), name = "my_pdf_attr"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
It can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{grouped}{A character string naming the grouping method:
\itemize{
\item \code{"col"} for grouping by colony
\item \code{"gen"} for grouping by generation
\item \code{"pop"} for no grouping (whole population)
}}

\item{groups}{The IDs of the groups for which to plot the PDF, a vector of positive integer values.
This argument is ignored in case \code{grouped = "pop"}.
The default value \code{-1} stands for all existing groups in the \code{tree}.}

\item{Ngroups}{Number of colonies in the movie (if \code{grouped = "col"}) or
number of generations in the movie (if \code{grouped = "gen"}), a non-zero positive integer value.
This argument is ignored in case \code{grouped = "pop"}.}

\item{model}{A character string naming the distribution model to be fitted:
\itemize{
\item \code{"norm"} is for fitting the \emph{Normal} distribution.
\item \code{"gamma"} is for fitting the \emph{Gamma} distribution.
\item \code{"lnorm"} is for fitting the \emph{Lognormal} distribution.
\item \code{"auto"} is for finding the best-fit distribution.
This is accomplished by fitting separately the \emph{Normal}, \emph{Gamma} and \emph{Lognormal} distribution.
The best-fit distribution is then chosen using the Bayesian Inference Criterion (BIC),
according to which the best model is the one with the lowest numeric BIC value.
}
Each model is fitted using the \emph{maximum likelyhood estimation (MLE)} method
provided by \code{\link[fitdistrplus]{fitdist}}.
\cr
Note that the \emph{Gamma} and \emph{Lognormal} distributions can be fitted to attributes with
non-zero positive numeric values.
Zero values are automatically replaced by value \code{1e-6}.
For negative values, an error is produced.}

\item{plot3D}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
a 3D or 2D plot will be generated, respectively.
When the default value \code{TRUE} is used, a common 3D plot for all groups specified in \code{groups} is generated.
When the value \code{FALSE} is used, a separate 2D plot for each group specified in \code{groups} is generated.
This argument is ignored (regarded as \code{FALSE}) in case \code{grouped = "pop"}
or if only one group is specified in \code{groups}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{250}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_pdf_attr"}.}
}}
}
\value{
A dataframe with the following columns:
\enumerate{
\item \code{group} is the ID of the group (a positive integer value)
or \code{-2} in case \code{grouped = "pop"}.
\item \code{Ncells} is the number of cells, a positive integer value.
\item \code{distr} is a character string naming the distribution model that was fitted:
\code{"norm"} for \emph{Normal},
\code{"gamma"} for \emph{Gamma} and
\code{"lnorm"} for \emph{Lognormal} distribution or
\code{NA} if no distribution was fitted (less than 2 unique values of \code{attr} exist).
\item \code{mean} is the \emph{\ifelse{html}{\out{mu}}{\eqn{\mu}}} parameter (\emph{mean}) of the \emph{Normal} distribution (a numeric value),
or \code{NA} in case \code{distr != "norm"}.
\item \code{sd} is the \emph{\ifelse{html}{\out{sigma}}{\eqn{\sigma}}} parameter (\emph{standard deviation}) of the \emph{Normal} distribution (a non-zero positive numeric value),
or \code{NA} in case \code{distr != "norm"}.
\item \code{shape} is the \emph{\ifelse{html}{\out{alpha}}{\eqn{\alpha}}} parameter (\emph{shape}) of the \emph{Gamma} distribution (a non-zero positive numeric value),
or \code{NA} in case \code{distr != "gamma"}.
\item \code{rate} is the \emph{\ifelse{html}{\out{beta}}{\eqn{\beta}}} parameter (\emph{rate}) of the \emph{Gamma} distribution (a non-zero positive numeric value),
or \code{NA} in case \code{distr != "gamma"}.
\item \code{meanlog} is the \emph{\ifelse{html}{\out{mu}}{\eqn{\mu}}} parameter of the \emph{Lognormal} distribution (a numeric value),
or \code{NA} in case \code{distr != "lnorm"}.
\item \code{sdlog} is the \emph{\ifelse{html}{\out{sigma}}{\eqn{\sigma}}} parameter of the \emph{Lognormal} distribution (a non-zero positive numeric value),
or \code{NA} in case \code{distr != "lnorm"}.
\item \code{BIC} is the BIC value of the fitted distribution (a numeric value),
or \code{NA} in case \code{distr = NA}.
\item \code{dBIC} is a character string summarizing the strength of the chosen distribution model
specified in \code{distr} against the other models with higher BIC values.
Value is \code{NA} in case \code{model != "auto"} or if \code{distr = NA}.
\cr\cr
The format of the string is \code{"<dBIC_norm>, <dBIC_gamma>, <dBIC_lnorm>"}.
Each \code{<dBIC_model>} value is rounded.
The larger a \code{<dBIC_model>} value, the stronger the evidence that attribute \code{attr} of the \code{group}
follows the chosen \code{distr} distribution against the \code{<model>} distribution.
Values \code{>10} typically indicate strong preference to the chosen distribution.
}
For groups with \code{distr = NA}, no plot is generated.
In case no cells exist, no plot is generated and \code{NULL} is returned.
}
\description{
Computes and plots the Probability Density Function (PDF) of a numeric attribute of a lineage or division tree.
The PDF can be plotted for each colony or generation or for the whole population.
Each PDF is computed by fitting a distribution model (\emph{Normal}, \emph{Gamma} or \emph{Lognormal})
to the corresponding data.
}
\details{
Each PDF is computed considering all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
\cr\cr
The range of x-axis (attribute) values depicted in each 2D plot is common and
is calculated as the range of values of all groups specified in \code{groups}.
\cr\cr
Color denotes the corresponding group.
}
