% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_attr.R
\name{plot_time_attr}
\alias{plot_time_attr}
\title{Plot mean and sd of an attribute per frame}
\usage{
plot_time_attr(tree, treeT = c("LT", "DT"), attr, unit = "",
  grouped = c("col", "gen", "pop"), groups = -1, Ngroups, save = FALSE,
  savePars = list(w = 2500, h = 2000, res = 350, path = getwd(), name =
  "my_time_attr"))
}
\arguments{
\item{tree}{The lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attr}{The name of the attribute in the \code{tree}, a character string.
In case \code{treeT = "LT"}, it can be any numeric attribute, as returned from \code{\link{get_attr_names}},
except for \code{"colony"}, \code{"generation"} and \code{"frame"}.
In case \code{treeT = "DT"}, it can be any numeric attribute, as returned from \code{\link{get_attr_names}},
having the suffix \code{"_birth"} or \code{"_division"}.}

\item{unit}{The unit of \code{attr}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attr} is in arbitrary units.}

\item{grouped}{A character string naming the grouping method:
\itemize{
\item \code{"col"} for grouping by colony
\item \code{"gen"} for grouping by generation
\item \code{"pop"} for no grouping (whole population)
}}

\item{groups}{The IDs of the groups for which to create the plot, a vector of positive integer values.
This argument is ignored in case \code{grouped = "pop"}.
The default value \code{-1} stands for all existing groups in the \code{tree}.}

\item{Ngroups}{Number of colonies in the movie (if \code{grouped = "col"}) or
number of generations in the movie (if \code{grouped = "gen"}), a non-zero positive integer value.
This argument is ignored in case \code{grouped = "pop"}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot(s)
will be saved in \code{.png} file(s) or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of each generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2500}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{2000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{350}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_time_attr"}.}
}}
}
\value{
A named list with the following components:
\item{group}{The ID of the group (a positive integer value)
or \code{-2} in case \code{grouped = "pop"}.}
\item{data}{A dataframe with the following columns:
\enumerate{
\item \code{frame} is the frame ID, a non-zero positive integer value.
\item \code{Ncells} is the number of cells, a positive integer value.
\item \code{mean} is the \emph{mean} of \code{attr}, a numeric value.
\item \code{sd} is the \emph{standard deviation} of \code{attr}, a positive numeric value
or \code{NA} in case \code{Ncells = 1}.
}
For groups with no cells, no plot is generated and \code{NULL} is returned.
}
}
\description{
Computes and plots the \emph{mean} and \emph{standard deviation} of a numeric attribute of a lineage or division tree
per frame. The plot can be created for each colony or generation or for the whole population.
}
\details{
A separate plot for each group specified in \code{groups} is generated.
Each plot is created considering all cells that are included in the analysis,
as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in \code{attr}.
\cr\cr
x-axis represents the time in \emph{frames}.
The range of x and y-axis values depicted in each plot is common and
is calculated as the range of the corresponding values of all specified groups.
\cr\cr
Color denotes the corresponding group.
}
