% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a tree}
\usage{
plot_tree(tree, treeT = c("LT", "DT"), attrC = "", unitC = "",
  NC = NULL, attrS = "", cellsC = "", colorCol = FALSE, Ncols,
  showLabels = FALSE, sizeLabel = 0.1, circular = TRUE,
  showLegends = TRUE, sizeL = 1, sizeV = 0.5, sizeE = 0.1,
  save = FALSE, savePars = list(w = 6000, h = 6000, res = 600, path =
  getwd(), name = "my_tree"))
}
\arguments{
\item{tree}{The connected lineage or division tree, an object of class \code{"igraph"}.}

\item{treeT}{A character string naming the type of \code{tree}:
\itemize{
\item \code{"LT"} if \code{tree} is a lineage tree
\item \code{"DT"} if \code{tree} is a division tree
}}

\item{attrC}{The name of the attribute in the \code{tree} by which the cells will be colored, a character string.
It can be any numeric or boolean attribute, as returned from \code{\link{get_attr_names}}.
Coloring is applied to all non-root cells, as returned from \code{\link{get_cells}},
except for cells with \code{NA} value in this attribute which are colored gray.
Any existing imaginary \emph{root} cells are always colored white.
When the default value \code{""} (the empty character) is used, nodes' color is the default.}

\item{unitC}{The unit of \code{attrC}, a character string.
It should be in the format \code{"<string>,<number>"},
where \code{",<number>"} represents the power and is optional
(e.g. \code{"m"} for meters and \code{"cm,3"} for cubic centimeters).
The default value is the empty character \code{""}, which implies that \code{attrC} is in arbitrary units.
This argument is ignored in case \code{attrC = ""}.}

\item{NC}{Number of colonies in the movie (if \code{attrC = "colony"}) or
number of generations in the movie (if \code{attrC = "generation"}), a non-zero positive integer value.
This argument is ignored in case \code{attrC != "colony"}
and \code{attrC != "generation"}.}

\item{attrS}{The name of the attribute in the \code{tree} by which the cells will be shaped, a character string.
It can be any boolean attribute, as returned from \code{\link{get_attr_names}}.
Cells are represented by dots or squares if their value in this attribute is \code{TRUE} or \code{FALSE}, respectively.
When the default value \code{""} (the empty character) is used, nodes' shape is the default.}

\item{cellsC}{The labels of the cells in the \code{tree} which will be colored red,
a vector of character strings.
They can be any cells, as returned from \code{\link{get_cells}}.
The default value is the empty character \code{""}, which stands for no cell.}

\item{colorCol}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
the edges of the tree will be colored based on the colony or not, respectively.
The value \code{TRUE} can be used only if \code{tree} contains the imaginary main \emph{root} cell.
When the default value \code{FALSE} is used, edges' color is the default.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.
This argument is ignored in case \code{colorCol = FALSE}.}

\item{showLabels}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
the labels of the cells will be shown on the plot or not, respectively.
The default value is \code{FALSE}.}

\item{sizeLabel}{Size of cell labels, a non-zero positive numeric value.
The default value is \code{0.1}.
This argument is ignored in case \code{showLabels = FALSE}.}

\item{circular}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
the tree will be plotted in circular or classical tree layout, respectively.
The default value is \code{FALSE}.}

\item{showLegends}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether
the explanatory legends and title will be shown on the plot or not, respectively.
The default value is \code{FALSE}.}

\item{sizeL}{Size of explanatory legends and title, a non-zero positive numeric value.
The default value is \code{1}.
This argument is ignored in case \code{showLegends = FALSE}.}

\item{sizeV}{Size of vertices, a non-zero positive numeric value.
The default value is \code{0.5}.}

\item{sizeE}{Width of edges, a non-zero positive numeric value.
The default value is \code{0.1}.}

\item{save}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the generated plot
will be saved in a \code{.png} file or displayed in the Plots Pane of RStudio, respectively.
The default value is \code{FALSE}.}

\item{savePars}{A named list specifying the parameters of the generated image file.
This argument is ignored in case \code{save = FALSE}.
Elements of the list are the following:
\describe{
\item{\code{w}}{The width of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{6000}.}
\item{\code{h}}{The height of the image file in \emph{pixels}, a non-zero positive integer value.
The default value is \code{6000}.}
\item{\code{res}}{The resolution of the image file in \emph{pixels} per \emph{inch} (ppi), a non-zero positive integer value.
The smaller this value, the larger the plot area in inches, and the smaller the text relative to the graph itself.
The default value is \code{600}.}
\item{\code{path}}{A character string naming the directory where the image file will be saved (excluding the last \code{"/"}).
If it does not contain an absolute path, the image file will be saved relative to the current working directory \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
\item{\code{name}}{The image file name, a character string.
The suffix \code{".png"} is added automatically.
The default value is \code{"my_tree"}.}
}}
}
\description{
Plots a lineage or division tree.
}
\details{
Nodes represent all cells, as returned from \code{\link{get_cells}}.
By default, all nodes are white dot points and edges are colored darkgray.
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected.
}
