% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tree.R
\name{save_tree}
\alias{save_tree}
\title{Save a tree on disc}
\usage{
save_tree(tree, savePath = getwd(), prefix = "", sep = "\\t")
}
\arguments{
\item{tree}{The linegae or division tree to be saved, an object of class \code{"igraph"}.}

\item{savePath}{A character string naming the directory where the \code{"nodes.csv"} and \code{"edges.csv"} files will be saved.
If it does not contain an absolute path, the files will be saved relative to the current working directory, \code{getwd()}.
The default value is the current working directory \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}

\item{prefix}{A prefix that will be added to the name of the \code{"nodes.csv"} and \code{"edges.csv"} files,
a character string.
The default value is the empty string \code{""}.}

\item{sep}{The field separator character.
Values on each line of the \code{"nodes.csv"} and \code{"edges.csv"} files will be separated by this character.
The default value is the tab separator \code{"\t"}.}
}
\description{
Saves a lineage or division tree on disc in nodes and edges \code{.csv} files.
}
