% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_subtree.R
\name{select_subtree}
\alias{select_subtree}
\title{Select a subtree}
\usage{
select_subtree(tree, criteria)
}
\arguments{
\item{tree}{The lineage or division tree from which to select a subtree, an object of class \code{"igraph"}.}

\item{criteria}{A list containing the selection criteria.
Each criterion (element of the list) should be a named list with the following components:
\describe{
\item{\code{attr}}{A character string naming the attribute in the \code{tree} on which the selection will be made.
It can be any numeric or boolean attribute, as returned from \code{\link{get_attr_names}}.}
\item{\code{val}}{The value of comparison.
It should be a numeric value when \code{attr} is numeric and
a logical value (\code{TRUE} or \code{FALSE}) when \code{attr} is boolean.}
\item{\code{op}}{The comparison operator.
It should be \code{"=="}, \code{"!="}, \code{"<"}, \code{"<="}, \code{">"} or \code{">="} when \code{attr} is numeric and
\code{"=="} or \code{"!="} when \code{attr} is boolean.}
}}
}
\value{
The selected subtree, an object of class \code{"igraph"}.
}
\description{
Selects a subtree of a lineage or division tree
based on multiple selection criteria combined with logical AND operator(s).
}
\details{
The selection is made on the nodes of the tree.
Only edges whose endpoints are both included in the resulting subtree are included.
Therefore, the resulting subtree may be disconnected.
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected,
\code{\link{unite_trees}} for combining multiple selection criteria with logical OR operator(s).
}
