% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_trees.R
\name{unite_trees}
\alias{unite_trees}
\title{Unite multiple trees}
\usage{
unite_trees(trees)
}
\arguments{
\item{trees}{A list with the trees to be united.
Each element of the list is a lineage or division tree (an object of class \code{"igraph"}).}
}
\value{
The united tree, an object of class \code{"igraph"}.
}
\description{
Unites multiple lineage or division trees from the same movie in one.
This function is useful for combining multiple selection criteria with logical OR operator(s).
}
\details{
Nodes and edges of the resulting tree are the union of nodes and edges of the trees, respectively.
Therefore, the resulting tree is disconnected.
Any missing attributes among the trees are filled with \code{NA} values in the resulting tree.
}
\seealso{
\code{\link{select_subtree}} for combining multiple selection criteria with logical AND operator(s).
}
