% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cell.R
\name{view_cell}
\alias{view_cell}
\title{View a cell}
\usage{
view_cell(LT, cells, cell_list, col_list, Ncols)
}
\arguments{
\item{LT}{The lineage tree where the cells specified in \code{cells} belong, an object of class \code{"igraph"}.}

\item{cells}{The labels of the cells in the \code{LT} to be viewed, a vector of character strings.
They can be any non-root cells, as returned from \code{\link{get_cells}}.}

\item{cell_list}{A list containing all the cell instants of the movie.}

\item{col_list}{A list containing all the colony instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}
}
\description{
Displays the images of a set of given cells in the Plots Pane of RStudio.
}
\details{
A separate image for each cell specified in \code{cells} is generated.
The cell is viewed in its colony and is marked as red.
The rest cells of the colony are marked as white.
}
\section{Prerequisites}{

This function can be used by \emph{BaSCA} or \emph{SuperSegger} users,
importing the data with \code{\link{import_basca}} or \code{\link{import_ss}}, respectively.
\cr\cr
If \code{\link{import_json}} was used for importing the data,
it is necessary that cell list elements have the \code{pixelList} and \code{colId} components
and colony list elements have the \code{colImage} component.
See \code{\link{import_json}} for more details.
In other case, this function cannot be used (throws an error).
\cr\cr
Users of \emph{Oufti} who imported the data with \code{\link{import_oufti}}
are also \bold{excluded} from using this function, as no colony list was returned.
}

