% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFDT.R
\name{createFDT}
\alias{createFDT}
\title{Create cells' FDT}
\usage{
createFDT(LTmain, minLife = 5, frameR)
}
\arguments{
\item{LTmain}{The main part of the overall FLT,
a connected lineage tree containing the imaginary \emph{root} cells
(object of class \code{"igraph"}).}

\item{minLife}{Minimum life in \emph{frames} for a cell to be included in the analysis, a positive integer value.
The default value is \code{5}. Use value \code{0} to exclude all leaf cells from the analysis.}

\item{frameR}{Frame rate of the movie in \emph{frames} per \emph{minute}, a non-zero positive numeric value.}
}
\value{
A named list with the following components:
\item{DTmain}{The corresponding main part of the overall FDT,
a connected division tree containing the imaginary \emph{root} cells (object of class \code{"igraph"}).}
\item{LTmain}{The updated LTmain with the attributes \code{"generation"} and \code{"age"} updated/added,
an object of class \code{"igraph"}.
The \code{"age"} attribute denotes the age of each cell instant in \emph{frames}.}
\item{Ngens}{Number of generations in the movie, a non-zero positive integer value.
IDs of generations are in the range \code{[0, Ngens-1]}.}
}
\description{
Creates the cells' Forest of Division Trees (FDT) given the corresponding Forest of Lineage Trees (FLT).
}
\details{
A continuous segment (sequence) of LT nodes between two successive cell \emph{divisions} represents the lifespan of a cell.
A cell is \emph{divided} if
\itemize{
\item it gives birth to 2 daughter cells in the next frame,
\item it disappears from the field of view in the next frame or
\item it is not linked to any cell instant in the next frame due to tracking errors.
}
The fuction creates the Forest of Division Tree (FDT) by reducing all LT cell segments down to single nodes (DT nodes),
except for the imaginary \emph{root} cells.
\cr\cr
Each node of the FDT represents a cell at its full lifespan, having as character string attributes
the concatenation of all attributes of the \emph{collapsed} LT nodes by \code{", "},
with the following exceptions:
\itemize{
\item The \code{"colony"} attribute is not concatenated,
since the ID of the colony from which the cell emanated characterizes both cell and its instants.
The \code{"colony"} attribute is again a non-zero positive integer number.
\item Attributes \code{"age"} and/or \code{"generation"} are also not concatenated in case they exist,
since they are (re)evaluated by the function and updated/stored in the FLT.
See below for more details.
\item Each boolean attribute in the FLT forms also a boolean attribute in the FDT,
with value (\code{TRUE} or \code{FALSE}) based on the majority vote of the corresponding values
of the \emph{collapsed} LT nodes.
}
The concatenated \code{"name"} attribute represents the labels of the \emph{collapsed} cells (instants of the cell)
and is renamed to \code{"cellInstants"}.
The \code{"name"} attribute is again a non-zero positive integer number stored as a character string,
denoting the label of the cell in the FDT.
Value \code{"1"} corresponds to the main \emph{root} cell.
Values \code{"1+<i>"} correspond to the colonies' \emph{root} cells, where \code{"<i>"} is the colony ID.
The rest values correspond to the cells.
\cr\cr
For each numeric attribute in the FLT
(except for \code{"colony"} and \code{"frame"}, plus \code{"age"} and/or \code{"generation"} in case they exist),
the concatenation represents the cell's time-series of the attribute.
Given each cell's time-series of an attribute \code{"<attr>"},
the following numeric life attributes are estimated and stored as attributes in the corresponding FDT node:
\itemize{
\item \code{"<attr>_birth"} is the \code{"<attr>"} value of the first instant of the cell
\item \code{"<attr>_division"} is the \code{"<attr>"} value of the last instant of the cell
\item \code{"<attr>_mean"} is the \emph{mean} of \code{"<attr>"}
\item \code{"<attr>_sd"} is the \emph{standard deviation} of \code{"<attr>"},
or \code{NA} in case the cell has only one instant
\item \code{"<attr>_min"} is the minimum value of \code{"<attr>"}
\item \code{"<attr>_max"} is the maximum value of \code{"<attr>"}
}
These numeric attributes are in units of \code{"<attr>"}.
\cr\cr
The following attributes are also life attributes and are stored in each DT node:
\itemize{
\item \code{"generation"} is the ID of the generation of the cell, a positive integer value.
This value is also updated/stored in the \code{"generation"} attribute of the instants of the cell in the FLT,
since the ID of the generation characterizes both cell and its instants.
\item \code{"birthTime"} is the ID of the frame at which the cell is firstly spotted (born),
a non-zero positive integer value.
This value is the \code{"frame"} value of the first instant of the cell.
\item \code{"divisionTime"} is the ID of the frame at which the cell is lastly spotted (a frame before its \emph{division}),
a non-zero positive integer value.
This value is the \code{"frame"} value of the last instant of the cell.
\item \code{"lifeFrames"} is the duration of the cell life in \emph{frames}, a non-zero positive integer value.
This value is computed as
\code{\ifelse{html}{\out{lifeFrames = divisionTime - birthTime + 1}}{\deqn{lifeFrames = divisionTime - birthTime + 1}}}.
\item \code{"lifeHours"} is the duration of the cell life in \emph{hours}, a non-zero positive numeric value.
This value is computed as
\code{\ifelse{html}{\out{lifeHours = lifeFrames / (60 * frameR)}}{\deqn{lifeHours = \frac{lifeFrames}{60 \cdot frameR}}}}.
\item \code{"isConsidered"} is a logical value (\code{TRUE} or \code{FALSE})
indicating whether the cell will be included in the analysis or not.
This value is \code{FALSE} for the imaginary \emph{root} cells.
The value for the cells is computed based on the \code{minLife} argument.
It is \code{FALSE} for all leaf cells and \code{TRUE} for the rest, in case \code{minLife = 0}, or
\code{FALSE} for all leaf cells with \code{"lifeFrames" <= minLife} and \code{TRUE} for the rest, in case \code{minLife != 0}.
}
}
\seealso{
\code{\link{isConnected}} for checking if a tree is connected,
\code{\link{save_tree}} for saving a tree on disc.
}
