% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFLT.R
\name{createFLT}
\alias{createFLT}
\title{Create cells' FLT}
\usage{
createFLT(cell_list, Ncols)
}
\arguments{
\item{cell_list}{A list containing all the cell instants of the movie.}

\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.}
}
\value{
A named list with the following components:
\item{LTmain}{The main part of the overall FLT,
a connected lineage tree containing the imaginary \emph{root} cells (object of class \code{"igraph"}).}
\item{branches}{A list with the motherless branches of the overall FLT.
Each branch (element of the list) is a connected lineage tree (object of class \code{"igraph"}).
Motherless branches arise from tracking errors, in case a cell instant (root of the branch) fails to be
connected to any cell instant of the previous frame, or when a cell instant (root of the branch) just entered the field of view.}
}
\description{
Creates the cells' Forest of Lineage Trees (FLT) given a list containing all the cell instants of the movie.
}
\details{
Obligatory attributes are:
\itemize{
\item \code{"name"} is the ID of the cell, a non-zero positive integer number stored as a character string.
Value \code{"1"} corresponds to the main \emph{root} cell.
Values \code{"1+<i>"} correspond to the colonies' \emph{root} cells,
where \code{"<i>"} is the colony ID.
\item \code{"cellName"} is the name of the cell, a character string.
The main \emph{root} cell is named as \code{"root"}.
The colonies' \emph{root} cells are named as \code{"colony<i>"},
where \code{"<i>"} is the colony ID.
\item \code{"frame"} is the ID of the frame of the cell, a non-zero positive integer number
\item \code{"colony"} is the ID of the colony of the cell, a non-zero positive integer number
}

Any other numeric or boolean attribute of a cell is also stored as node attribute to the FLT.
The imaginary \emph{root} cells have value \code{-1} in all numeric attributes and value \code{FALSE} in all boolean attributes.

Apart from the cell instants of the movie,
an imaginary main \emph{root} cell as well as imaginary \emph{root} cells for each colony are also added to the FLT.
Colonies' \emph{root} cells are daughters of the main \emph{root} cell.
Cell instants of the first frame of the movie are daughters of the corresponding colony's \emph{root} cell.
The imaginary \emph{root} cells are used to facilitate the tree representation of the movie and the colony tracking
and are automatically excluded from the analysis.
}
\seealso{
\code{\link{save_tree}} for saving a tree on disc,
\code{\link{add_branch}} for connecting a motherless branch to a lineage tree.
}
