% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_json.R
\name{import_json}
\alias{import_json}
\title{Import custom-made cell and colony list}
\usage{
import_json(file_cells, file_cols = NULL)
}
\arguments{
\item{file_cells, file_cols}{Character strings naming the \code{.json} files
(including the suffix \code{".json"}) containing all the cell and colony instants of the movie, respectively,
from which the data is to be imported.
If a string does not contain an absolute path, the file name is relative to the current working directory, \code{getwd()}.
\cr\cr
When the default value \code{file_cols = NULL} is used, no colony list is imported.
\cr\cr
NOTE: The components should be separated by \code{/} (not \code{\\}) on Windows.}
}
\value{
A named list with the following components:
\item{cell_list}{A list containing all the cell instants of the movie.
Each element of the list is a named list having as components the corresponding key-value pairs in the \code{file_cells}.}
\item{cell_list}{A list containing all the colony instants of the movie.
Each element of the list is a named list having as components the corresponding key-value pairs in the \code{file_cols}.}
\item{Nframes}{Number of frames in the movie, a non-zero positive integer value.
IDs of frames are in the range \code{[1, Nframes]}.}
\item{Ncols}{Number of colonies in the movie, a non-zero positive integer value.
IDs of colonies are in the range \code{[1, Ncols]}.
This value corresponds to the number of colonies at the start of the movie.}
}
\description{
Imports a \code{.json} file containing all the cell instants of the movie
and another \code{.json} file containing all the colony instants of the movie.
}
\details{
The \code{file_cells} must be a JSON array of objects.
Each object denotes a cell instant of the movie and must contain the following key-value pairs:
\itemize{
\item \code{cellName} is the name of the cell, a character string in the format \code{"<cell>_f<frame>"}.
\item \code{frame} is the ID of the frame of the cell, a non-zero positive integer number.
\item \code{colony} is the ID of the colony of the cell in the \code{frame},
a non-zero positive integer number.
If the whole frame is treated as a single colony, value \code{1} must be used for all cells.
\item \code{daughterIds} is a JSON array of character strings containing the \code{cellName}
of the linked cell(s) in the next frame.
In case one such cell exists, it can also be a character string.
In case no such cells exist, it can either be \code{NULL} or omitted.
\item \code{colId} is a pointer to the corresponding colony instant of the cell in the \code{file_cols},
a non-zero positive integer value.
\cr\cr
NOTE: This key-value pair should be omitted if \code{file_cols = NULL}. A warning message is produced.

\item \code{pixelList} is a JSON array of arrays.
Each element of the array is a JSON array of 2 non-zero integer values,
indicating a pixel of the cell in colony coordinates
(i.e. relative to the \code{colImage} of the \code{colId} element
in the colony list returned from \code{\link{import_col_list}}).
The first value represents the row and the second the column of the pixel.
\cr\cr
NOTE: This key-value pair should be omitted if \code{file_cols = NULL} or
if objects in \code{file_cols} do not contain the key \code{colImage}.
A warning message is produced.
}

\itemize{
\item \code{colName} is the name of the colony instant, a character string in the format \code{"f<frame>_c<colony>"},
where \code{"<frame>"} and \code{"<colony>"} is the ID of the frame and colony (in the frame) of the colony instant, respectively.

\item \code{prev_colName} is a vector of character strings containing the \code{colName}
of the corresponding colony instant(s) in the previous frame.
For colony instants of the first frame this is equal to \code{"f0_c0"}.

\item \code{next_colName} is a vector of character strings containing the \code{colName}
of the corresponding colony instant(s) in the next frame.
For colony instants of the last frame this is equal to \code{"f00_c0"}.

\item \code{colImage} is the mask of the box surrounding the colony instant, a matrix of \code{0} and \code{1}.
\code{1}s denote the pixels of cells and \code{0}s the background pixels.

\item \code{ULcorner} is a vector of 2 non-zero integer values
denoting the upper-left pixel of the box surrounding the colony instant in global (frame) coordinates.
The first integer represents the row and the second the column of the pixel.
}
}
