% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ss.R
\name{import_ss}
\alias{import_ss}
\title{Import \emph{SuperSegger} data}
\usage{
import_ss(path = "/home/vic/Desktop/sal15_9_ss", somebasename = "sal15_9",
  xy_number = "", Ndigits = 3, Nframes, pixelR)
}
\arguments{
\item{pixelR}{The pixel ratio in units of length, a non-zero positive numeric value.}

\item{file}{A character string naming the \code{.csv} file generated by \emph{Oufti}
(including the suffix \code{".csv"}) from which the data is to be imported.
If it does not contain an absolute path, the file name is relative to the current working directory, \code{getwd()}.
\cr\cr
NOTE: The components should be separated by \code{"/"} on Windows.}
}
\value{
A named list with the following components:
\item{col_list}{A list containing all the colony instants of the movie.
Each element of the list is a named list with the following components:
\itemize{
\item \code{colName} is the name of the colony instant, a character string in the format \code{"f<frame>_c1"},
where \code{"<frame>"} is the ID of the frame of the colony instant.
\item \code{prev_colName} is a vector of character strings containing the \code{colName}
of the corresponding colony instant(s) in the previous frame.
For colony instants of the first frame, this is equal to \code{"f0_c0"}.
\item \code{next_colName} is a character string containing the \code{colName}
of the corresponding colony instant in the next frame.
For colony instants of the last frame, this is equal to \code{"f00_c0"}.
\item \code{colImage} is the mask of the box surrounding the colony instant, a matrix of \code{0} and \code{1}.
\code{1}s denote the pixels of cells and \code{0}s the background pixels.
\item \code{ULcorner} is an 1x2 matrix of non-zero integer values
denoting the upper-left pixel of the box surrounding the colony instant in global (frame) coordinates.
The first integer represents the row and the second the column of the pixel.
}
}

\item{cell_list}{A list containing all the cell instants of the movie.
Each element of the list is a named list with the following components:
\itemize{
\item \code{cellName} is the name of the cell, a character string in the format \code{"c<cellId>_f<frame>"}.
\item \code{frame} is the ID of the frame of the cell, a non-zero positive integer number.
\item \code{colony} is the ID of the colony of the cell in the \code{frame}, a non-zero positive integer number.
\item \code{daughterIds} is a vector of character strings containing the \code{cellName}
of the linked cell(s) in the next frame,
or \code{NULL} in case no such cells exist.
\item \code{colId} is a \emph{pointer} to the corresponding colony instant of the cell in the \code{col_list},
a non-zero positive integer value.
\cr\cr
Colonies that entered the field of view at a time point and did not exist from the beginning of the movie (i.e. from the first frame)
should not have tracked cells, until they merge (if this is the case) with another existing colony.
This means that no element should \emph{point} to such colony instants.
\item \code{pixelList} is a Nx2 matrix of non-zero integer values
denoting the pixels of the cell in colony coordinates
(i.e. relative to the \code{colImage} of the \code{colId}^th element in the \code{col_list}).
Each one of the N rows indicates a pixel of the cell.
The first column represents the row and the second the column of the pixel.
\item \code{centroid} is an 1x2 matrix of non-zero numeric values
denoting the centroid (geometric center) of the cell in colony coordinates
(i.e. relative to the \code{colImage} of the \code{colId}^th element in the \code{col_list}).
It is the \emph{mean} of the \code{pixelList} by column.
\item \code{length} is the length of the cell in units of length, a non-zero positive numeric value.
\item \code{width} is the width of the cell in units of length, a non-zero positive numeric value.
\item \code{LW} is the length-to-width ratio, a non-zero positive numeric value.
\item \code{area} is the area of the cell in squared units of length, a non-zero positive numeric value.
\item \code{perimeter} is the perimeter of the cell in units of length, a non-zero positive numeric value.
\item \code{eccentricity} is a numeric value in the range \code{[0, 1]}
defining the eccentricity of the cell.
\item \code{orientation} is a numeric value in the range \code{[0, 360)}
defining the orientation of the cell in \emph{degrees}.
\item \code{solidity} is a numeric value in the range \code{[0, 1]}
defining the solidity of the cell.
\item \code{fluorescenceInt<i>} is a Nx1 matrix of positive numeric values indicating the
fluoresence intensity of channel \code{<i>} for each one of the pixels in \code{pixelList}
(1-1 correspondence).
\item \code{fluorescenceInt<i>.mean} is the \emph{mean} of \code{fluorescenceInt<i>}, a positive numeric value.
\item \code{fluorescenceInt<i>.std} is the \emph{standard deviation} of \code{fluorescenceInt<i>}, a positive numeric value.
\item \code{fluorescenceInt<i>.coverage} is the percentage of the cell area covered by fluoresence,
a numeric value in the range \code{[0, 1]}.
}
}

\item{Nframes}{Number of frames in the movie, a non-zero positive integer value.
IDs of frames are in the range \code{[1, Nframes]}.}
\item{Ncols}{Number of colonies in the movie, equal to \code{1}.
IDs of colonies are in the range \code{[1, Ncols]}.}
\item{frameH}{Frame image height in \emph{pixels}, a non-zero positive integer value.}
\item{frameW}{Frame image width in \emph{pixels}, a non-zero positive integer value.}
}
\description{
Imports a \code{.csv} file generated by \emph{SuperSegger} and converts them into
a cell list and colony list containing all the cell instants and colony instants of the movie, respectively.
}
\details{
\emph{SuperSegger} treats the whole frame as a single colony.
\cr\cr
\emph{SuperSegger} users importing the data with this function are \bold{excluded} from using
\code{\link{split_cell}}, \code{\link{get_cand_merge_cells}} and \code{\link{merge_cells}}.
}
\references{
\url{http://mtshasta.phys.washington.edu/website/SuperSegger.php}
}
